+++ src/EyeRest/Properties/Resources.Designer.cs
@@ -69,5 +69,15 @@ namespace EyeRest.Properties {
                 return ((System.Drawing.Icon)(obj));
             }
         }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Icon similar to (Icon).
+        /// </summary>
+        internal static System.Drawing.Icon AppSnoozeIcon {
+            get {
+                object obj = ResourceManager.GetObject("AppSnoozeIcon", resourceCulture);
+                return ((System.Drawing.Icon)(obj));
+            }
+        }
     }
 }